+++ packages/haiku-timeline/src/components/RowSegments.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import mixpanel from 'haiku-serialization/src/utils/Mixpanel'
 import TransitionBody from './TransitionBody'
 import ConstantBody from './ConstantBody'
 import SoloKeyframe from './SoloKeyframe'
@@ -124,7 +125,20 @@ export default class RowSegments extends React.Component {
             <div
               id={`keyframe-container-${keyframe.getUniqueKey()}`}
               key={`keyframe-container-${keyframe.getUniqueKey()}`}
-              className={`keyframe-container no-select`}>
+              className={`keyframe-container no-select`}
+              onDoubleClick={(doubleClickEvent) => {
+                if (
+                  doubleClickEvent.target &&
+                  doubleClickEvent.target.id &&
+                  doubleClickEvent.target.id.includes('keyframe-dragger')
+                ) {
+                  this.props.timeline.seekToTime(keyframe.origMs)
+                  this.props.row.blurOthers({ from: 'timeline' })
+                  this.props.row.focus({ from: 'timeline' })
+                  this.props.row.select({ from: 'timeline' })
+                  mixpanel.haikuTrack('creator:timeline:keyframe:double-clicked')
+                }
+              }}>
               {segmentPieces}
             </div>
           )